+++ packages/core/src/index.ts
@@ -19,7 +19,7 @@ export {
 } from "./install.js";
 export { identifyInstall, scanForInstalls } from "./discovery.js";
 export { backupDirectory, debugDirectory, logFile, packageDirectory, temporaryDirectory } from "./directories.js";
-export { appendLog } from "./log.js";
+export { appendLog, type LogLevel } from "./log.js";
 export { fnv1a } from "./hash.js";
 export { copyTree, listFilesRecursively } from "./fs.js";
 export { mergeIni, type MergeConflict } from "./ini-merge.js";